Added an fps counter and random seeding to color changing

+++ src/Functions.c
@@ -1,6 +1,7 @@
 #include "include/Functions.h"
 #include <stdlib.h>
 
+
 float randomColor() {
-    return (float)(0.1f * (rand() % 11));
+    return (float)(0.1f * (float)(rand() % 11));
 }
